modif h

+++ src/main/java/fr/koumare/comptease/model/Article.java
@@ -23,10 +23,10 @@ public class Article {
     @ElementCollection
     private List<String> category;
 
-    private Long quantite;
+    private int quantite;
 
     @Column(name = "price", nullable = false, precision = 10, scale = 2)
-    private BigDecimal price;
+    private Double price;
 
     @ManyToMany(mappedBy = "articles")
     private List<Devis> devis;
@@ -37,13 +37,13 @@ public class Article {
     public Article() {
     }
 
-    public Article(String description, List<String> category, Long quantite, BigDecimal price) {
+    public Article(String description, List<String> category, int quantite, Double price) {
         this.description = description;
         this.category = category;
         this.quantite = quantite;
         this.price = price;
     }
-    public Long getQuantite() {
+    public int getQuantite() {
         return quantite;
     }
 }
